% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{data}
\name{Activation_gatingTemplate}
\alias{Activation_gatingTemplate}
\title{CytoRSuite Activation gatingTemplate.}
\format{a \code{gatingTemplate} object containing the following 9 gates:
\itemize{ \item{\strong{Cells}} \item{\strong{Single Cells}}
\item{\strong{Live Cells}} \item{\strong{T Cells}} \item{\strong{Dendritic
Cells}} \item{\strong{CD4 T Cells}} \item{\strong{CD8 T Cells}}
\item{\strong{CD69+ CD4 T Cells}} \item{\strong{CD69+ CD8 T Cells}} }}
\source{
The Activation_gatingTemplate was generated using CytoRSuite and the
Activation dataset.
}
\description{
CytoRSuite gatingTemplate used to gate the Activation dataset.
}
\seealso{
\link{CytoRSuiteData}

\link{Compensation}

\link{Activation}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
