% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{data}
\name{Compensation_gatingTemplate}
\alias{Compensation_gatingTemplate}
\title{CytoRSuite gatingTemplate used gate the Compensation dataset.}
\format{a \code{gatingTemplate} containing the following 2 gates: \itemize{
\item{\strong{Cells}} \item{\strong{Single Cells}}}}
\source{
Some compensation controls from an experiment.
}
\description{
A gatingTemplate used to gate the Compensation dataset prior to
spillover calculation.
}
\seealso{
\link{CytoRSuiteData}

\link{Compensation}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
