% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoExploreRData.R
\docType{package}
\name{CytoExploreRData}
\alias{CytoExploreRData}
\alias{CytoExploreRData-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
CytoExploreRData
A data package for CytoExploreRData.

Use \code{data(package='CytoExploreRData')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{Compensation_gatingTemplate}
}
