% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{package}
\name{CytoRSuiteData}
\alias{CytoRSuiteData}
\alias{CytoRSuiteData-package}
\title{Flow Cytometry Data for cytoRSuite}
\description{
Contains flow cytometry data for use in cytoRSuite.
}
\details{
CytoRSuiteData A data package for cytoRSuite.

Use \code{data(package='CytoRSuiteData')$results[, 3]} to see a list
of available data sets in this data package and/or
DataPackageR::load_all_datasets() to load them.
}
\seealso{
\link{example_gatingTemplate}
}
