% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{data}
\name{example_gatingTemplate}
\alias{example_gatingTemplate}
\title{Example gatingTemplate used to gate the Activation dataset}
\format{a \code{gatingTemplate} containing the following gates: \describe{
\item{Cells} \item{Single Cells} \item{Live Cells} \item{T Cells}
\item{Dendritic Cells} \item{CD4 T Cells} \item{CD8 T Cells} \item{CD69+
CD4 T Cells} \item{CD69+ CD8 T Cells} }}
\source{
The example_gatingTemplate was generated using cytoRSuite and the
Activation dataset.
}
\description{
Example gatingTemplate used to gate the Activation dataset.
}
\seealso{
\link{CytoRSuiteData}
}
