% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{data}
\name{Activation}
\alias{Activation}
\title{Descriptive data title}
\format{a \code{flowSet} containing the following fields:
\describe{
}}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{CytoRSuiteData}
}
