% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{data}
\name{Activation}
\alias{Activation}
\title{CytoRSuite Activation Dataset}
\format{a \code{flowSet} containing 4 samples in order of increasing OVA
concentration.}
\source{
The data comes from an in vitro OT-I/II T cell activation assay.
}
\description{
Data from an in vitro OT-I/II T cell activation assay.
}
\details{
\describe{ An experiment performed to assess the in vitro activation
of OT-I and OT-II transgenic T cells co-cultured with OVA-presenting
dendritic cells (DCs). This dataset is comprised of 4 samples saved as a
\code{flowSet} object. Samples are arranged in order of increasing OVA
concentration.

Pooled DCs & T cells were labelled as follows: \itemize{ \item{\strong{Hoechst
33258} a DNA intercalating dye which fluoresces in Alexa Fluor 405-A and
Alexa Fluor 430-A channels.} \item{\strong{CD11c-APC-Cy7} to identify DCs within the
samples.} \item{\strong{Va2-PE} to identify cells which express the OVA-specific TCR
(i.e. OT-I and OT-II T cells).} \item{\strong{CD8-FITC} to identify the CD8+ OT-I
transgenic T cells (Alexa Fluor 488-A).} \item{\strong{CD4-Alexa Fluor 700} to
identify the CD4+ OT-II transgenic T cells (Alexa Fluor 700-A).}
\item{\strong{CD44-APC} to assess the extent of T cell activation (Alexa Fluor 647-A).}
\item{\strong{CD69-PerCP Cy5.5} to assess the extent of T cell activation (7-AAD-A).} }}
}
\seealso{
\link{CytoRSuiteData}

\link{Compensation}

\link{Activation_gatingTemplate}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
