% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoExploreRData.R, R/CytoRSuiteData.R
\docType{data}
\name{Activation_gatingTemplate}
\alias{Activation_gatingTemplate}
\title{CytoExploreR Activation gatingTemplate.}
\format{a \code{gatingTemplate} object containing the following 9 gates:
\itemize{ \item{\strong{Cells}} \item{\strong{Single Cells}}
\item{\strong{Live Cells}} \item{\strong{T Cells}} \item{\strong{Dendritic
Cells}} \item{\strong{CD4 T Cells}} \item{\strong{CD8 T Cells}}
\item{\strong{CD69+ CD4 T Cells}} \item{\strong{CD69+ CD8 T Cells}} }}
\source{
The Activation_gatingTemplate was generated using CytoExploreR to
analyse Activation data set.

The Activation_gatingTemplate was generated using CytoRSuite to
analyse Activation data set.
}
\description{
CytoExploreR gatingTemplate used to gate the Activation data set.

CytoRSuite gatingTemplate used to gate the Activation data set.
}
\seealso{
\link{CytoExploreRData}

\link{Compensation}

\link{Activation}

\link{CytoRSuiteData}

\link{Compensation}

\link{Activation}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)

Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
