% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{data}
\name{Compensation}
\alias{Compensation}
\title{A collection of single stain compensation controls and an unstained
control used in cytoRSuite.}
\format{a \code{flowSet} containing the following 6 compensation controls:
\describe{
\item{7AAD.fcs}
\item{AF-430.fcs}
\item{APC-Cy7.fcs}
\item{NIL.fcs}
\item{PE-Cy7.fcs}
\item{PE.fcs}
}}
\source{
Some compensation controls from an experiment.
}
\description{
Compensation control samples.
}
\seealso{
\link{CytoRSuiteData}
}
