% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{data}
\name{Compensation}
\alias{Compensation}
\title{CytoRSuite Compensation Dataset.}
\format{A \code{flowSet} containing the 6 compensation controls.}
\source{
Some compensation controls from an experiment.
}
\description{
A collection of single stain compensation controls and an
unstained control used in CytoRSuite for testing and demonstration
purposes.
}
\details{
Compensation controls are as follows: \itemize{
\item{\strong{7AAD.fcs}} \item{\strong{AF-430.fcs}}
\item{\strong{APC-Cy7.fcs}} \item{\strong{NIL.fcs}{  (Unstained)}}
\item{\strong{PE-Cy7.fcs}} \item{\strong{PE.fcs}} }
}
\seealso{
\link{CytoRSuiteData}

\link{Activation}

\link{Activation_gatingTemplate}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
