% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoExploreRData.R, R/CytoRSuiteData.R
\docType{data}
\name{Compensation}
\alias{Compensation}
\title{CytoExploreR Compensation Data Set.}
\format{A \code{flowSet} containing the 7 compensation controls.}
\source{
Compensation controls used for an in vitro OT-I/II T cell activation
assay.

Compensation controls used for an in vitro OT-I/II T cell activation
assay.
}
\description{
A collection of single stain compensation controls and an
unstained control used to compensate the Activation data set.

A collection of single stain compensation controls and an
unstained control used to compensate the Activation data set.
}
\details{
Compensation controls are as follows: \itemize{
\item{\strong{Compensation-7AAD.fcs}}
\item{\strong{Compensation-AF700.fcs}} \item{\strong{Compensation-APC.fcs}}
\item{\strong{Compensation-APC-Cy7.fcs}}
\item{\strong{Compensation-FITC.fcs}} \item{\strong{Compensation-PE.fcs}}
\item{\strong{Compensation-Unstained.fcs}}}

Compensation controls are as follows: \itemize{
\item{\strong{Compensation-7AAD.fcs}}
\item{\strong{Compensation-AF700.fcs}} \item{\strong{Compensation-APC.fcs}}
\item{\strong{Compensation-APC-Cy7.fcs}}
\item{\strong{Compensation-FITC.fcs}} \item{\strong{Compensation-PE.fcs}}
\item{\strong{Compensation-Unstained.fcs}}}
}
\seealso{
\link{CytoExploreRData}

\link{Activation}

\link{Activation_gatingTemplate}

\link{CytoRSuiteData}

\link{Activation}

\link{Activation_gatingTemplate}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)

Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
