% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoExploreRData.R, R/CytoRSuiteData.R
\docType{data}
\name{Compensation_gatingTemplate}
\alias{Compensation_gatingTemplate}
\title{CytoExploreR gatingTemplate used gate the Compensation data set.}
\format{a \code{gatingTemplate} containing the following 3 gates: \itemize{
\item{\strong{nonDebris}} \item{\strong{Cells}} \ item{\strong{Single
Cells}}}}
\source{
The compensation gatingTemplate was generated using CytoExploreR to
gate the compensation controls prior to claculation of fluorescent
spillover.

The compensation gatingTemplate was generated using CytoRSuite to
gate the compensation controls prior to claculation of fluorescent
spillover.
}
\description{
A gatingTemplate used to gate the Compensation data set prior to
spillover calculation.

A gatingTemplate used to gate the Compensation data set prior to
spillover calculation.
}
\seealso{
\link{CytoExploreRData}

\link{Compensation}

\link{CytoRSuiteData}

\link{Compensation}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)

Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
