% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{package}
\name{CytoRSuiteData}
\alias{CytoRSuiteData}
\alias{CytoRSuiteData-package}
\title{CytoRSuiteData}
\description{
Contains flow cytometry datasets for use in CytoRSuite,
including a flowSet of compensation controls, an experimental flowSet
containing 4 samples and an example gatingTemplate.
}
\details{
A data package for CytoRSuite.

\itemize{ \item{\strong{Compensation}}{ A flowSet of 6 compensation
controls, including 5 named single stain controls and a universal unstained
control (NIL.fcs). These compensation controls are not associated with the
Activation dataset and are used for testing and demonstration purposes
only. Compensation controls include: \itemize{\item{\strong{7AAD.fcs}}
\item{\strong{APC-Cy7.fcs}} \item{\strong{NIL.fcs}}{  (Unstained)}
\item{\strong{PE-Cy7.fcs}} \item{\strong{PE.fcs}}} }
\item{\strong{Activation}}{ A flowSet of 4 samples from an in vitro OT-I/II
activation assay. In this assay dendritic cells (DCs) were activated and
pulsed with increasing concentrations of ovalbumin. Following overnight
culture, activated DCs were co-cultured with OT-I/II OVA-specific T cells
overnight. The extent of T activation was assessed by flow cytometry as
follows: \itemize{ \item{\strong{Hoechst 33258}}{  a DNA intercalating dye
which fluoresces in Alexa Fluor 405-A and Alexa Fluor 430-A channels.}
\item{\strong{CD11c-APC-Cy7}}{  to identify DCs within the samples.}
\item{\strong{Va2-PE}}{  to identify cells which express the OVA-specific
TCR (i.e. OT-I and OT-II T cells).} \item{\strong{CD8-FITC}}{ to identify
the CD8+ OT-I transgenic T cells (Alexa Fluor 488-A).}
\item{\strong{CD4-Alexa Fluor 700}}{  to identify the CD4+ OT-II transgenic
T cells (Alexa Fluor 700-A).} \item{\strong{CD44-APC}}{  to assess the
extent of T cell activation (Alexa Fluor 647-A).} \item{\strong{CD69-PerCP
Cy5.5}}{  to assess the extent of T cell activation (7-AAD-A).} } }
\item{\strong{Activation_gatingTemplate}{ an example gatingTemplate
generated by CytoRSuite to gate the Activation dataset. This gatingTemplate
includes the following gates: \itemize{ \item{\strong{Cells}}
\item{\strong{Single Cells}} \item{\strong{Live Cells}} \item{\strong{T
Cells}} \item{\strong{Dendritic Cells}} \item{\strong{CD8 T Cells}}
\item{\strong{CD4 T Cells}} \item{\strong{CD69+ CD4 T Cells}}
\item{\strong{CD69+ CD8 T Cells}} }}}}
}
\seealso{
\link{Compensation}

\link{Activation}

\link{Activation_gatingTemplate}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
