% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoRSuiteData.R
\docType{package}
\name{CytoRSuiteData}
\alias{CytoRSuiteData}
\alias{CytoRSuiteData-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
CytoRSuiteData
A data package for CytoRSuiteData.

Use \code{data(package='CytoRSuiteData')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{Activation}
}
